Fix all TypeScript errors
The errors were because components that are rendered directly through
react-router have to declare their state and props arguments both as
type `any`.

This might not be a good long term idea. But right now, I'm learning
both React and TypeScript and can't have tons of errors spitting out
because it makes it difficult to see real errors that I'm introducing.

See:
* https://github.com/ReactTraining/react-router/issues/4317
* https://github.com/brunolm/ts-react-redux-startup/issues/2